Extract the ModelBuilder and reuse
