+++ test/java/org/opendatakit/briefcase/export/ModelTest.java
@@ -23,9 +23,7 @@ import static org.hamcrest.Matchers.is;
 import static org.javarosa.core.model.instance.TreeReference.DEFAULT_MULTIPLICITY;
 import static org.junit.Assert.assertThat;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.javarosa.core.model.DataType;
@@ -64,50 +62,6 @@ public class ModelTest {
         .build();
 
     assertThat(model.getChoices(), empty());
-  }
-
-  static class ModelBuilder {
-    private TreeElement current = new TreeElement(null, DEFAULT_MULTIPLICITY);
-    private Map<String, QuestionDef> controls = new HashMap<>();
-
-    ModelBuilder addGroup(String name) {
-      TreeElement child = new TreeElement(name, DEFAULT_MULTIPLICITY);
-      child.setDataType(DataType.NULL.value);
-      child.setRepeatable(false);
-      child.setParent(current);
-      current.addChild(child);
-      current = child;
-      return this;
-    }
-
-    ModelBuilder addRepeatGroup(String name) {
-      TreeElement child = new TreeElement(name, DEFAULT_MULTIPLICITY);
-      child.setDataType(DataType.NULL.value);
-      child.setRepeatable(true);
-      child.setParent(current);
-      current.addChild(child);
-      current = child;
-      return this;
-    }
-
-    ModelBuilder addField(String name, DataType dataType) {
-      return addField(name, dataType, null);
-    }
-
-    ModelBuilder addField(String name, DataType dataType, QuestionDef control) {
-      TreeElement child = new TreeElement(name, DEFAULT_MULTIPLICITY);
-      child.setDataType(dataType.value);
-      child.setParent(current);
-      current.addChild(child);
-      current = child;
-      if (control != null)
-        controls.put(Model.fqn(current, 0), control);
-      return this;
-    }
-
-    Model build() {
-      return new Model(current, controls);
-    }
   }
 
   @Test
@@ -149,4 +103,5 @@ public class ModelTest {
       child = child.children().get(0);
     return child;
   }
+
 }
